refactor: extract indexer/discover.rs (phase 3, step 9)
Move find_source_files, walkdir_find, find_source_files_unconstrained,
find_source_files_newer_than, warm_discover_files out of indexer.rs
into src/indexer/discover.rs. All functions are pure (no self/Indexer
dependency). indexer.rs imports them via use self::discover::{...}.

Add discover_tests.rs with 7 tests:
- empty dir returns no files
- .kt / .java files discovered
- IgnoreMatcher exclusion respected
- unconstrained scan includes build/ dirs
- warm_discover_files returns cached on-disk files
- warm_discover_files skips deleted files
